+++ optimus/lib/src/group_wrapper.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/widgets.dart';
-import 'package:optimus/optimus.dart';
-import 'package:optimus/src/theme/theme.dart';
-import 'package:optimus/src/theme/theme_data.dart';
+import 'package:optimus/src/common/field_error.dart';
+import 'package:optimus/src/common/field_label.dart';
 
 class GroupWrapper extends StatelessWidget {
   const GroupWrapper({
@@ -18,32 +17,18 @@ class GroupWrapper extends StatelessWidget {
   final bool isRequired;
 
   @override
-  Widget build(BuildContext context) {
-    final theme = OptimusTheme.of(context);
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        if (label != null && label!.isNotEmpty)
-          OptimusLabelSmall(
-            child: Text(
-              isRequired ? '${label!} *' : label!,
-              style: TextStyle(color: _textColor(theme)),
+  Widget build(BuildContext context) => Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          if (label != null && label!.isNotEmpty)
+            OptimusFieldLabel(
+              label: label!,
+              isRequired: isRequired,
             ),
-          ),
-        child,
-        if (error != null && error!.isNotEmpty)
-          OptimusCaption(
-            child: Text(
-              error!,
-              style: TextStyle(color: theme.colors.danger500),
-            ),
-          ),
-      ],
-    );
-  }
-
-  Color _textColor(OptimusThemeData theme) =>
-      theme.isDark ? theme.colors.neutral0t64 : theme.colors.neutral1000t64;
+          child,
+          if (error != null && error!.isNotEmpty)
+            OptimusFieldError(error: error!),
+        ],
+      );
 }
